+++ include/QtMate/widgets/headerViewSortSectionFilter.hpp
@@ -22,6 +22,7 @@
 #include <QHeaderView>
 #include <QEvent>
 #include <QSet>
+#include <QMouseEvent>
 
 namespace qtMate
 {